backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.10.1
# Azure SDK for AI services
openai==1.52.0
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0
azure-cognitiveservices-speech==1.40.0
# Additional utilities for RAG
aiofiles==23.2.1
# Azure AI Search and Blob Storage
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
# Document processing
pypdf2==3.0.1
python-docx==1.1.0
# RAG and document processing
pathlib-mate==1.3.2  # Path utilities